Verifiy authenticity of the majority of downloads (#20)
Use GPG and SHA256 for verification, unless said verification resources are unavailable.

In the case where National Instruments does not offer a way to verify the signature, the SHA256 is compared with the Packages file they provide. While the Packages file has a signature, it is unusable as the public key is not online so there is no verification of the Packages file, so it will continue to be reliant on HTTPS.

The only two files that cannot be verified are cloog and isl, as they don't provide signatures nor are HTTPS secured.